+++ app/src/main/java/com/developerspace/webrtcsample/compose/ui/screens/ChatListScreen.kt
@@ -27,7 +27,7 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.lifecycle.viewmodel.compose.viewModel
-import com.developerspace.webrtcsample.ChatMainActivity
+import com.developerspace.webrtcsample.legacy.ChatMainActivity
 import com.developerspace.webrtcsample.compose.ui.theming.MyTheme
 import com.developerspace.webrtcsample.compose.ui.theming.lightBlue
 import com.developerspace.webrtcsample.compose.ui.viewmodel.ChatListViewModel
