+++ app/src/main/java/com/developerspace/webrtcsample/compose/ui/screens/UserDetailScreen.kt
@@ -46,7 +46,7 @@ import com.developerspace.webrtcsample.compose.ui.theming.lightGreen
 import com.developerspace.webrtcsample.compose.ui.util.AppLevelCache
 import com.developerspace.webrtcsample.compose.ui.viewmodel.UserDetailViewModel
 import com.developerspace.webrtcsample.model.User
-import com.developerspace.webrtcsample.util.MyOpenDocumentContract
+import com.developerspace.webrtcsample.util.misc.MyOpenDocumentContract
 import com.google.firebase.auth.ktx.auth
 import com.google.firebase.ktx.Firebase
 import kotlin.math.sqrt
