+++ app/src/main/java/com/developerspace/webrtcsample/legacy/ChatMainActivity.kt
@@ -1,4 +1,4 @@
-package com.developerspace.webrtcsample
+package com.developerspace.webrtcsample.legacy
 
 import android.content.Context
 import android.net.Uri
@@ -11,8 +11,9 @@ import androidx.appcompat.app.AppCompatActivity
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import androidx.recyclerview.widget.RecyclerView.Recycler
+import com.developerspace.webrtcsample.R
 import com.developerspace.webrtcsample.model.FriendlyMessage
-import com.developerspace.webrtcsample.util.MyOpenDocumentContract
+import com.developerspace.webrtcsample.util.misc.MyOpenDocumentContract
 import com.firebase.ui.database.FirebaseRecyclerOptions
 import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.auth.ktx.auth
