+++ app/src/main/java/com/developerspace/webrtcsample/legacy/Constants.kt
@@ -1,4 +1,4 @@
-package com.developerspace.webrtcsample
+package com.developerspace.webrtcsample.legacy
 
 class Constants {
     companion object {
