+++ app/src/main/java/com/developerspace/webrtcsample/legacy/MyScrollToBottomObserver.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.developerspace.webrtcsample
+package com.developerspace.webrtcsample.legacy
 
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
