+++ app/src/main/res/layout/active_users.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activeUsers.ui.ActiveUserActivity">
+    tools:context=".legacy.activeUsers.ui.ActiveUserActivity">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/recyclerView"
