+++ app/src/main/res/layout/activity_main.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:keepScreenOn="true"
-    tools:context=".RTCActivity">
+    tools:context=".legacy.RTCActivity">
 
     <org.webrtc.SurfaceViewRenderer
         android:id="@+id/remote_view"
