+++ app/src/main/res/layout/activity_sign_in.xml
@@ -4,6 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SignInActivity">
+    tools:context=".legacy.SignInActivity">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
