+++ app/src/main/res/layout/activity_start.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_margin="16dp"
     android:layout_height="match_parent"
-    tools:context=".MainActivity"
+    tools:context=".legacy.MainActivity"
     android:orientation="vertical">
 
     <TextView
